merge from master

+++ common/item_struct.h
@@ -83,7 +83,7 @@ struct Item_Struct {
 	char	Lore[80];		// Lore Name: *=lore, &=summoned, #=artifact, ~=pending lore
 	char	IDFile[30];		// Visible model
 	uint32	ID;				// Unique ID (also PK for DB)
-	uint8	Weight;			// Item weight * 10
+	int32	Weight;			// Item weight * 10
 	uint8	NoRent;			// No Rent: 0=norent, 255=not norent
 	uint8	NoDrop;			// No Drop: 0=nodrop, 255=not nodrop
 	uint8	Size;			// Size: 0=tiny, 1=small, 2=medium, 3=large, 4=giant
@@ -160,7 +160,7 @@ struct Item_Struct {
 	//uint32	Unk059;
 	union {
 		uint32 Fulfilment;	// Food fulfilment (How long it lasts)
-		int16 CastTime;		// Cast Time for clicky effects, in milliseconds
+		uint32 CastTime;		// Cast Time for clicky effects, in milliseconds
 	};
 	uint32 EliteMaterial;
 	int32	ProcRate;
